backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
sentence-transformers
faiss-cpu
python-multipart
pypdf
aiofiles
